update get products by category and get all products pages / calls
