+++ src/app/_components/productTile.tsx
@@ -51,7 +51,7 @@ export default function ProductTile(props: ProductTileProps) {
   const { setOpacity, setType, setOperation } = useContext(BannerContext);
   return (
     <div
-      className={`bg-${appTheme}-containerBg text-${appTheme}-text border-${appTheme}-border border-2 rounded-lg p-4 h-96`}
+      className={`bg-${appTheme}-containerBg text-${appTheme}-text border-${appTheme}-border border-2 rounded-lg p-4 h-96 cursor-pointer`}
     >
       <Link
         href={
