+++ src/app/products/page.tsx
@@ -1,4 +1,3 @@
-import { getProducts } from "../_utils/serverutils";
 import {
   ListHeader,
   SearchBar,
@@ -14,12 +13,6 @@ export default async function AllProducts({
   searchParams?: { query?: string };
 }) {
   const query = searchParams?.query || "";
-  const products = await getProducts();
-  const filteredProducts = Array.isArray(products)
-    ? products.filter((product) => {
-        return product.name.toLowerCase().includes(query.toLowerCase());
-      })
-    : [];
 
   return (
     <PageWrapper>
@@ -31,11 +24,7 @@ export default async function AllProducts({
         </div>
       </div>
       <Suspense fallback={<LoadingSpinner />}>
-        <ProductContainer
-          products={filteredProducts.sort(
-            (a, b) => Number(a.categoryId) - Number(b.categoryId)
-          )}
-        />
+        <ProductContainer query={query} />
       </Suspense>
     </PageWrapper>
   );
